requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3
PyPDF2==3.0.1
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4